functions/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-cloud-firestore==2.13.1
google-api-python-client==2.108.0
google-auth==2.23.4
firebase-admin==6.2.0
functions-framework==3.4.0
